requirements.txt
python-dotenv
vk_api
flask
requests
sqlalchemy
psycopg2-binary
flask-async